Fix typescript error with adapter
