+++ packages/start/vite/index.d.ts
@@ -1,11 +1,5 @@
 import { Plugin } from "node_modules/vite";
 import { Options } from "./plugin";
 
-export type Adapter = {
-  start(options: Options): Promise<void>;
-  build(options: Options): Promise<void>;
-  dev(options: Options): Promise<void>;
-  name: string;
-};
-
+export { Adapter } from './plugin';
 export default function (opts?: Partial<Options>): Plugin[];
